ENH: refs #0379. Don't allow public, private, or base folders to be deleted
-They can't be deleted using the batch delete now.

+++ core/controllers/BrowseController.php
@@ -379,7 +379,8 @@ class BrowseController extends AppController
         continue;
         }
 
-      if($this->Folder->policyCheck($folder, $this->userSession->Dao, MIDAS_POLICY_ADMIN))
+      if($this->Folder->policyCheck($folder, $this->userSession->Dao, MIDAS_POLICY_ADMIN) &&
+         $this->Folder->isDeleteable($folder))
         {
         $this->Folder->delete($folder);
         $resp['success']['folders'][] = $folderId;
